# back-end/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Education website with Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "name@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "api/freeDocumentCategory/createFreeDocumentCategory/": {
            "post": {
                "description": "Return free document category's informaion.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Document category name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "api/freeDocument/destroyFreeDocument/{id}": {
            "post": {
                "description": "Return freeDocument's informaion.",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "integer"
                                },
                                "file": {
                                    "type": "string"
                                },
                                "category_id": {
                                    "type": "integer"
                                },
                                "path": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "api/freeDocument/blockActiveFreeDocument/{id}": {
            "post": {
                "description": "Return freeDocument's informaion.",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "integer"
                                },
                                "file": {
                                    "type": "string"
                                },
                                "category_id": {
                                    "type": "integer"
                                },
                                "path": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "timestamp"
                                },
                                "updated_at": {
                                    "type": "timestamp"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}